Fix all the project structure in python and .NET
