+++ MecSolutionAccelerator.sln
@@ -13,7 +13,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Alerts.RulesEngine", "src\S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Alerts.Test", "src\Services\Alerts\Tests\Alerts.Test.csproj", "{E1F5F275-8CFF-47C2-B2B1-FD77330FB6B6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FilesManagement", "src\Files\FilesManagement\FilesManagement.csproj", "{F446DDC1-6C70-4847-B833-2F2072041552}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Files.Management", "src\Services\Files\Management\Files.Management.csproj", "{FC68E076-AD3B-4B06-8055-50B0AD0EFF28}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Files.MinIOInfraestructure", "src\Services\Files\MinIO\Files.MinIOInfraestructure.csproj", "{CA63DC21-1BAC-4F78-8421-55664D5D083F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Files.Contracts", "src\Services\Files\Contracts\Files.Contracts.csproj", "{F933CFE8-BB71-48B3-B5FD-7B8FF9D6FC9B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -41,10 +45,18 @@ Global
 		{E1F5F275-8CFF-47C2-B2B1-FD77330FB6B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E1F5F275-8CFF-47C2-B2B1-FD77330FB6B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E1F5F275-8CFF-47C2-B2B1-FD77330FB6B6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F446DDC1-6C70-4847-B833-2F2072041552}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F446DDC1-6C70-4847-B833-2F2072041552}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F446DDC1-6C70-4847-B833-2F2072041552}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F446DDC1-6C70-4847-B833-2F2072041552}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FC68E076-AD3B-4B06-8055-50B0AD0EFF28}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FC68E076-AD3B-4B06-8055-50B0AD0EFF28}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FC68E076-AD3B-4B06-8055-50B0AD0EFF28}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FC68E076-AD3B-4B06-8055-50B0AD0EFF28}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CA63DC21-1BAC-4F78-8421-55664D5D083F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA63DC21-1BAC-4F78-8421-55664D5D083F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA63DC21-1BAC-4F78-8421-55664D5D083F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA63DC21-1BAC-4F78-8421-55664D5D083F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F933CFE8-BB71-48B3-B5FD-7B8FF9D6FC9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F933CFE8-BB71-48B3-B5FD-7B8FF9D6FC9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F933CFE8-BB71-48B3-B5FD-7B8FF9D6FC9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F933CFE8-BB71-48B3-B5FD-7B8FF9D6FC9B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
