+++ docker-compose.override.yml
@@ -58,7 +58,7 @@ services:
       - ASPNETCORE_URLS=https://+:443;http://+:80
     ports:
       - "50052:80"
-      - "50058:443"
+      - "50078:443"
     volumes:
       - ${APPDATA}/Microsoft/UserSecrets:/root/.microsoft/usersecrets:ro
       - ${APPDATA}/ASP.NET/Https:/root/.aspnet/https:ro
