--- src/Files/FilesManagement/Commands/UploadNewFileCommand.cs
@@ -1,12 +1,0 @@
-﻿using MediatR;
-
-namespace Microsoft.MecSolutionAccelerator.Services.Files.Commands
-{
-    public class UploadNewFileCommand : IRequest<Guid>
-    {
-        public IFormFile FormFile { get; set; }
-        public string BucketName { get; set; }
-        public string SourceId { get; set; }
-        public long Timestamp { get; set; }
-    }
-}
