+++ src/Services/Detections/ai_inferencer/src/inference.py
@@ -7,63 +7,32 @@ import numpy as np
 import avro.schema
 from avro_json_serializer import AvroJsonSerializer
 import logging
-import boto3
-from botocore.client import Config
 import os
-
+import shared.minio_utils as minio
 
 
 def PublishEvent(pubsub_name: str, topic_name: str, data: json):
     with DaprClient() as client:
         resp = client.publish_event(pubsub_name=pubsub_name, topic_name=topic_name, data=data, data_content_type='application/json')
 
-def download_stream_from_minio(bucket_name, object_name, endpoint_url, access_key, secret_key):
-    s3 = boto3.client('s3',
-                      endpoint_url=endpoint_url,
-                      aws_access_key_id=access_key,
-                      aws_secret_access_key=secret_key,
-                      config=Config(signature_version='s3v4'),
-                      region_name='us-east-1')
-    
-    try:
-        response = s3.get_object(Bucket=bucket_name, Key=object_name)
-        streaming_body = response['Body']
-        
-        all_chunks = b''
-        while True:
-
-
-            chunk = streaming_body.read(1024)  # read only 1KB at a time
-            if not chunk:
-                break
-            all_chunks += chunk
-
-        return all_chunks  # This will be the entire file in bytes.
-        
-    except Exception as e:
-        logging.error(f"An error occurred: {e}")
-        return None
-
-
-
 def main(source_id,timestamp,model,image_id,detection_threshold,path,time_trace):
     timestamp_init=int(time.time()*1000)
     logging.basicConfig(level=logging.DEBUG)
     logging.info(source_id)
 
-    endpoint = os.getenv('MINIOURL')  # ej. 'http://localhost:9000'
-    access_key = 'minio'
-    secret_key = 'minio123'
+    endpoint = os.getenv('MINIOURL')
     bucket = 'images'
-
-    image_stream = download_stream_from_minio(bucket, image_id + ".jpg", endpoint, access_key, secret_key)
+    
+    minioClient = minio.MinIOClient(endpoint, 'minio', 'minio123')
+    image_stream = minioClient.download_stream(bucket, image_id + ".jpg",)
     if image_stream is not None:
         img = cv2.imdecode(np.frombuffer(image_stream, np.uint8), cv2.IMREAD_COLOR)
     else:
         logging.error('Failed to download image')
         return
     val_to_compare_resize,_,_=img.shape
-    
+    # dim = (720, 576)
+
     if val_to_compare_resize>576:
         logging.info(f'Resizing to print')
         dim = (720,576)
@@ -72,6 +41,8 @@ def main(source_id,timestamp,model,image_id,detection_threshold,path,time_trace)
         frame_to_bytes=frame_resized.tobytes()
         frame = base64.standard_b64encode(frame_to_bytes)
         frame = frame.decode()
+        image_id_str = str(image_id)
+        minioClient.upload_bytes(bucket, image_id_str+'.jpg', frame_to_bytes)
 
     data = { "SourceId":source_id,
     "UrlVideoEncoded": "1.0",
@@ -99,10 +70,10 @@ def main(source_id,timestamp,model,image_id,detection_threshold,path,time_trace)
             
             if list(detections["confidence"].values())[idx] > detection_threshold:
                 
-                xmin=list(detections["xmin"].values())[idx]
-                xmax=list(detections["xmax"].values())[idx]
-                ymin=list(detections["ymin"].values())[idx]
-                ymax=list(detections["ymax"].values())[idx]
+                xmin = list(detections["xmin"].values())[idx]
+                xmax = list(detections["xmax"].values())[idx]
+                ymin = list(detections["ymin"].values())[idx]
+                ymax = list(detections["ymax"].values())[idx]
                 BoundingBoxes.append({"x": xmin, "y":ymin})
                 BoundingBoxes.append({"x": xmin, "y":ymax})
                 BoundingBoxes.append({"x": xmax, "y":ymin})
