+++ src/Services/Detections/frameSplitter/Dockerfile
@@ -1,10 +1,11 @@
 FROM python:3.8.16-slim-bullseye
 
-COPY requirements.txt .
+COPY frameSplitter/requirements.txt .
 RUN pip install --disable-pip-version-check --no-cache-dir -U -r requirements.txt \
     && rm requirements.txt
 
-COPY ./Datasets ./Datasets
-COPY ./src ./src
+COPY ./frameSplitter/Datasets ./Datasets
+COPY ./frameSplitter/src ./src
+COPY shared ./src/shared
 
 CMD cd src && python3 frameSplitter.py
