+++ src/Services/Detections/frameSplitter/src/frameSplitter.py
@@ -8,10 +8,8 @@ import base64
 import os
 import logging
 import requests
-import boto3
-from botocore.exceptions import NoCredentialsError
-from botocore.client import Config
 import uuid
+import shared.minio_utils as minio
 
 class VideoCapture:
     def __init__(self, name):
@@ -49,44 +47,10 @@ class VideoCapture:
         self.t.join()
         self.cap.release()
 
-def bucket_exists(s3, bucket_name):
-    try:
-        s3.head_bucket(Bucket=bucket_name)
-        return True
-    except:
-        return False
-
-def create_bucket(s3, bucket_name):
-    s3.create_bucket(Bucket=bucket_name)
-    print(f"Bucket {bucket_name} created.")
-
-def upload_bytes_to_minio(bucket_name, object_name, data_bytes, endpoint_url, access_key, secret_key):
-    s3 = boto3.client(
-        's3',
-        endpoint_url=endpoint_url,
-        aws_access_key_id=access_key,
-        aws_secret_access_key=secret_key,
-        config=Config(signature_version='s3v4'),
-        region_name='us-east-1'
-    )
-    
-    if not bucket_exists(s3, bucket_name):
-        create_bucket(s3, bucket_name)
-
-    try:
-        s3.put_object(Body=data_bytes, Bucket=bucket_name, Key=object_name)
-        print(f"Data has been uploaded to {bucket_name} as {object_name}.")
-    except NoCredentialsError:
-        print("Credentials not available.")
-    except Exception as e:
-        print(f"An error occurred: {e}")
-
 def main():
-    endpoint = os.getenv('MINIOURL')  # ej. 'http://localhost:9000'
-    access_key = 'minio'
-    secret_key = 'minio123'
+    endpoint = os.getenv('MINIOURL')
     bucket = 'images'
-    
+    minioClient = minio.MinIOClient(endpoint, 'minio', 'minio123')
     logging.basicConfig(level=logging.DEBUG)
     
     timer=0
@@ -94,9 +58,7 @@ def main():
     try:
         MY_POD_NAME=(os.getenv('MY_POD_NAME'))
         FEEDS=(os.getenv('FEEDS'))
-
         FEEDS=list(eval(FEEDS))
-
         dict_pos=MY_POD_NAME.split('-')[-1]
         feeds_dict=FEEDS[int(dict_pos)]
 
@@ -109,8 +71,6 @@ def main():
     logging.info(f'feed url: {feed_URL}')
         
     time.sleep(timer)
-    
- 
     cap = VideoCapture(feed_URL)
     while True:
         # Capture frame-by-frame
@@ -132,7 +92,7 @@ def main():
             image_id = uuid.uuid4()
             image_id_str = str(image_id)
             logging.info(f'Image uploaded with ID: {image_id}')
-            upload_bytes_to_minio(bucket, image_id_str+'.jpg', resized_img_bytes, endpoint, access_key, secret_key)
+            minioClient.upload_bytes(bucket, image_id_str+'.jpg', resized_img_bytes)
             with DaprClient() as client:
             # Create data to send to AI inference service
                 time_trace = {"stepStart": timestamp_init, "stepEnd": int(time.time() * 1000), "stepName": "frameSplitter"}
@@ -147,10 +107,7 @@ def main():
                 response = resp.text()
                 logging.info(response)
         except Exception as e:
-            logging.error(f'Error encountered: {e}')
-
-
-        
+            logging.error(f'Error encountered: {e}')     
 
 if __name__ == '__main__':
     main()
