+++ src/Services/Files/Management/Commands/DeleteFilesCommand.cs
@@ -4,6 +4,6 @@ namespace Microsoft.MecSolutionAccelerator.Services.Files.Commands
 {
     public class DeleteFilesCommand : IRequest
     {
-        public string bucketName { get; set; }
+        public string containerName { get; set; }
     }
 }
