+++ src/Services/Files/Management/Commands/DownloadFileCommand.cs
@@ -5,7 +5,7 @@ namespace Microsoft.MecSolutionAccelerator.Services.Files.Commands
     public class DownloadFileCommand : IRequest<DownloadFileResponse>
     {
         public string FileName { get; set; }
-        public string BucketName { get; set; }
+        public string ContainerName { get; set; }
     }
 
     public class DownloadFileResponse
