+++ src/Services/Files/Management/Dockerfile
@@ -5,23 +5,23 @@ FROM --platform=$BUILDPLATFORM mcr.microsoft.com/dotnet/sdk:7.0.400 AS build
 ARG TARGETARCH=amd64
 
 WORKDIR /src
-COPY ["src/Files/FilesManagement/FilesManagement.csproj", "src/Files/FilesManagement/"]
+COPY ["src/Services/Files/Management/Files.Management.csproj", "src/Services/Files/Management/"]
 
 # Restaurar considerando la arquitectura objetivo
-RUN dotnet restore "src/Files/FilesManagement/FilesManagement.csproj" -r linux-$TARGETARCH
+RUN dotnet restore "src/Services/Files/Management/Files.Management.csproj" -r linux-$TARGETARCH
 
 COPY . .
-WORKDIR "/src/src/Files/FilesManagement"
+WORKDIR "/src/src/Services/Files/Management"
 
 # Compilar teniendo en cuenta la arquitectura objetivo
-RUN dotnet build "FilesManagement.csproj" -c Release -r linux-$TARGETARCH -o /app/build
+RUN dotnet build "Files.Management.csproj" -c Release -r linux-$TARGETARCH -o /app/build
 
 FROM build AS publish
-RUN dotnet publish "FilesManagement.csproj" -c Release -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "Files.Management.csproj" -c Release -o /app/publish /p:UseAppHost=false
 
 # Imagen final basada en el runtime de .NET
 FROM mcr.microsoft.com/dotnet/aspnet:7.0.4 AS final
 WORKDIR /app
 EXPOSE 80
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "FilesManagement.dll"]
+ENTRYPOINT ["dotnet", "Files.Management.dll"]
