+++ src/Services/Files/Management/Files.Management.csproj
@@ -18,4 +18,9 @@
 	<PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Contracts\Files.Contracts.csproj" />
+    <ProjectReference Include="..\MinIO\Files.MinIOInfraestructure.csproj" />
+  </ItemGroup>
+
 </Project>
